Fix model loading and dtype errors
Issues fixed:
1. EfficientNet hash validation error
   - Migrated from deprecated pretrained=True to new weights API
   - Uses EfficientNet_B0_Weights.IMAGENET1K_V1
   - Added graceful fallback if weights download fails

2. MobileNet deprecation warning
   - Updated to use MobileNet_V2_Weights.IMAGENET1K_V1
   - Consistent with new torchvision API

3. Hybrid model ResNet loading
   - Updated to use ResNet50_Weights.IMAGENET1K_V1
   - Added error handling for weight loading failures

4. dtype mismatch error (expected Double but found Float)
   - Ensured image tensor is explicitly float32 in preprocess_image
   - Added .float() call to prevent dtype conversion issues

All models now use the modern torchvision weights API which fixes
hash validation errors and eliminates deprecation warnings.

+++ crop_monitoring_app/backend/utils.py
@@ -112,8 +112,8 @@ def preprocess_image(image_path, target_size=(224, 224)):
     std = np.array([0.229, 0.224, 0.225])
     image_array = (image_array - mean) / std
 
-    # Convert to tensor (C, H, W)
-    image_tensor = torch.from_numpy(image_array).permute(2, 0, 1).unsqueeze(0)
+    # Convert to tensor (C, H, W) and ensure float32 dtype
+    image_tensor = torch.from_numpy(image_array).permute(2, 0, 1).unsqueeze(0).float()
 
     return image_tensor, original_image
